Punt on unicode handling for now
I can't figure out what's up with the encoding in the HTML and how
to get it to properly write things to both the console and when
redirected to a file.  For now, just convert characters to ASCII,
replacing characters it doesn't know how to deal with.

Add a saved version of the offending HTML to the test data.  We
should test this in the future.